events_processor_service/requirements.txt
pika
sqlalchemy
pymysql
python-dotenv
requests
apscheduler